:sparkles:  add http gateway and change serviceid type from ushort to int32
新增http网关功能，修复服务端拦截器的类型错误，新增http网关的示例，修改序列化方案中必须固定CodecType字段

+++ src/DotBPE.Extra.Castle/InvokeMeta.cs
@@ -7,7 +7,7 @@ namespace DotBPE.Extra
 {
     public class InvokeMeta
     {
-        public ushort ServiceId { get; set; }
+        public int ServiceId { get; set; }
         public ushort MessageId { get; set; }
 
         public string ServiceGroupName { get; set; }
